+++ src/components/App.js
@@ -1,3 +1,5 @@
+import '../css/app.css'
+
 function App() {
   return <h1>The future is dark_</h1>
 }
